Strip out variable declaration logic from grammar
This will allow me to develop expressions separately from variable
declarations and then try to marry the two together later on. Given
that some expressions can be present in variable declarations I might
focus on this first. Expressions also make up the "meat" of a program
so there is a lot of work to be done here.