Fix most issues with line continuations
End-of-statements are now handled in the external scanner, which
allows us to use a bit more logic than the DSL rules allow.

Logic of external scanner now goes like:

1. skip all leading whitespace
2. work out if we should end the current statement
3. work out if we should end the current line continuation
4. handle number literals
5. possibly start a new line continuation

Statements are always ended by semicolons and end-of-file, and _may_
be ended by newlines unless a line continuation has started.

Because comments always go to the end of the line, and may appear
inside line continuations, it turns out to be easier to treat comments
as also ending statements -- unless they appear inside a line
continuation.

Line continuations now are essentially always in pairs marking the
start and end of the continuation, possibly with comments in-between.