Add parent node for procedure bindings
Previously it was hard to iterate over all the methods in a
`procedure_statement` as procedure aliases were at the same level.

This is a similar change to variable declarations, where adding a
new parent node made it easier to iterate over init declarations.

Also making this a field for consistency with variable declarations