Fix `unary_expression` precedence
Fortran standard has unary operator below multiplication and above
addition. Fixing this is the difference between

    -1**2

being (incorrectly) parsed as:

    (math_expression
      (unary_expression
        (number_literal))
      (number_literal))

and (correctly) parsed as:

    (unary_expression
      (math_expression
        (number_literal)
        (number_literal)))

I had to also fix `data_value` because the `unary_expression` was
trying to eat the final `/`. `data_value` only allows constants, so we
need to parse signed literals.